+++ tests/test_utils.py
@@ -1,0 +1,9 @@
+import nacl.utils
+
+
+def test_random_bytes_produces():
+    assert len(nacl.utils.random(16)) == 16
+
+
+def test_random_bytes_produces_different_bytes():
+    assert nacl.utils.random(16) != nacl.utils.random(16)
